test(codex): speed up Git Bash installer fixtures
